+++ app/src/main/java/com/example/newsapp/features/model/NewsResponse.kt
@@ -1,4 +1,4 @@
-package com.example.newsapp.model
+package com.example.newsapp.features.model
 
 data class NewsResponse(
     val articles: List<Article>,
