+++ app/src/main/java/com/example/newsapp/features/model/Source.kt
@@ -1,4 +1,4 @@
-package com.example.newsapp.model
+package com.example.newsapp.features.model
 
 data class Source(
     val id: String,
