+++ app/src/main/java/com/example/newsapp/features/repository/NewsRepository.kt
@@ -1,7 +1,7 @@
-package com.example.newsapp.repository
+package com.example.newsapp.features.repository
 
-import com.example.newsapp.di.NewsApiService
-import com.example.newsapp.model.NewsResponse
+import com.example.newsapp.network.NewsApiService
+import com.example.newsapp.features.model.NewsResponse
 import javax.inject.Inject
 
 class NewsRepository @Inject constructor(private val apiService: NewsApiService) {
