+++ app/src/main/java/com/example/newsapp/utils/Constants.kt
@@ -1,0 +1,4 @@
+package com.example.newsapp.utils
+
+object Constants {
+}
