+++ app/src/main/java/com/example/newsapp/utils/NetworkResponse.kt
@@ -1,0 +1,7 @@
+package com.example.newsapp.utils
+
+sealed class NetworkResponse<out T : Any?> {
+    object Loading : NetworkResponse<Nothing>()
+    data class Error(val errorMessage: String) : NetworkResponse<Nothing>()
+    data class Success<out T : Any>(val result: T?) : NetworkResponse<T>()
+}
